Align some of the 4.0.6a packets. Ported over addon loading. Cleaned up some shit code. Character screen should load now.

+++ src/server/game/Entities/Player/Player.cpp
@@ -24107,7 +24107,7 @@ void Player::AddKnownCurrency(uint32 itemId)
         //SetFlag64(PLAYER_FIELD_KNOWN_CURRENCIES, (1LL << (ctEntry->BitIndex-1)));
 }
 
-void Player::UpdateFallInformationIfNeed(MovementInfo const& minfo, uint16 opcode)
+void Player::UpdateFallInformationIfNeed(MovementInfo const& minfo, uint32 opcode)
 {
     if (m_lastFallTime >= minfo.fallTime || m_lastFallZ <= minfo.pos.GetPositionZ() || opcode == MSG_MOVE_FALL_LAND)
         SetFallInformation(minfo.fallTime, minfo.pos.GetPositionZ());
