+++ src/server/game/Entities/Player/Player.h
@@ -2293,7 +2293,7 @@ class Player : public Unit, public GridObject<Player>
         /*********************************************************/
         /***                 VARIOUS SYSTEMS                   ***/
         /*********************************************************/
-        void UpdateFallInformationIfNeed(MovementInfo const& minfo, uint16 opcode);
+        void UpdateFallInformationIfNeed(MovementInfo const& minfo, uint32 opcode);
         Unit *m_mover;
         WorldObject *m_seer;
         void SetFallInformation(uint32 time, float z)
